I DONT HAVE BRAINNNN
I'm not even sure if I'm going to finish it or not, I'm just writing this damn thing to get new experiences, I'm changing it, I don't know what I'm doing, I'm just pressing the keys on the damn keyboard and I guess that's not enough. Obviously this is a suicide note thank you to everyone I love for being a part of my life. @Surge13z don't forget me cute human